# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'Gemdy Calendar API'
  version: 1.0.0
  description: 'API REST para la gestión de cursos, pendientes y recordatorios de Gemdy Calendar.'
servers:
  -
    url: 'http://localhost:8000/api'
paths:
...
  '/cursos/{curso}':
    get:
      operationId: cursos.show
      description: 'Obtiene la información de un curso específico.'
      summary: 'Mostrar curso'
      tags:
        - Curso
      parameters:
        -
          name: curso
          in: path
          required: true
          description: 'The curso ID'
          schema:
            type: integer
      responses:
        200:
          description: '`CursoResource`'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CursoResource'
                required:
                  - data
        404:
          $ref: '#/components/responses/ModelNotFoundException'
    put:
      operationId: cursos.update
      description: 'Actualiza los datos de un curso existente.'
      summary: 'Actualizar curso'
      tags:
        - Curso
      parameters:
        -
          name: curso
          in: path
          required: true
          description: 'The curso ID'
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCursoRequest'
      responses:
        200:
          description: '`CursoResource`'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CursoResource'
                required:
                  - data
        404:
          $ref: '#/components/responses/ModelNotFoundException'
        422:
          $ref: '#/components/responses/ValidationException'
    delete:
      operationId: cursos.destroy
      description: 'Elimina un curso siempre que no tenga pendientes activos.'
      summary: 'Eliminar curso'
      tags:
        - Curso
      parameters:
        -
          name: curso
          in: path
          required: true
          description: 'The curso ID'
          schema:
            type: integer
      responses:
        204:
          description: 'No content'
        404:
          $ref: '#/components/responses/ModelNotFoundException'
        409:
          description: 'El curso tiene pendientes activos.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
...
  /pendientes/con-recordatorio:
    post:
      operationId: pendiente.storeConRecordatorio
      description: "Crea un pendiente junto con su recordatorio asociado.\n\nEste método se conserva como parte de la lógica desarrollada\nanteriormente, aunque no se expone actualmente mediante una ruta."
      summary: 'Crear pendiente con recordatorio'
      tags:
        - Pendiente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorePendienteConRecordatorioRequest'
      responses:
        201:
          description: '`PendienteResource`'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PendienteResource'
                required:
                  - data
        422:
          $ref: '#/components/responses/ValidationException'
...
  '/pendientes/{pendiente}':
    get:
      operationId: pendientes.show
      description: 'Obtiene la información de un pendiente específico.'
      summary: 'Mostrar pendiente'
      tags:
        - Pendiente
      parameters:
        -
          name: pendiente
          in: path
          required: true
          description: 'The pendiente ID'
          schema:
            type: integer
      responses:
        200:
          description: '`PendienteResource`'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PendienteResource'
                required:
                  - data
        404:
          $ref: '#/components/responses/ModelNotFoundException'
    put:
      operationId: pendientes.update
      description: 'Actualiza los datos de un pendiente existente.'
      summary: 'Actualizar pendiente'
      tags:
        - Pendiente
      parameters:
        -
          name: pendiente
          in: path
          required: true
          description: 'The pendiente ID'
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePendienteRequest'
      responses:
        200:
          description: '`PendienteResource`'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PendienteResource'
                required:
                  - data
        404:
          $ref: '#/components/responses/ModelNotFoundException'
        422:
          $ref: '#/components/responses/ValidationException'
        409:
          description: 'No se puede actualizar el pendiente porque incumple una regla de negocio.'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
    delete:
      operationId: pendientes.destroy
      description: 'Elimina un pendiente existente.'
      summary: 'Eliminar pendiente'
      tags:
        - Pendiente
      parameters:
        -
          name: pendiente
          in: path
          required: true
          description: 'The pendiente ID'
          schema:
            type: integer
      responses:
        204:
          description: 'No content'
        404:
          $ref: '#/components/responses/ModelNotFoundException'
...
components:
  schemas:
    CursoResource:
      type: object
      properties:
        id:
          type: integer
          description: 'Identificador del curso.'
          examples:
            - 1
        usuario_id:
          type: integer
          description: 'Identificador del usuario propietario.'
          examples:
            - 1
        nombre:
          type: string
          description: 'Nombre del curso.'
          examples:
            - 'Desarrollo de Software'
        codigo:
          type: string
          description: 'Código del curso.'
          examples:
            - IF-4101
        semestre:
          type: string
          description: 'Semestre del curso.'
          examples:
            - II-2026
        creditos:
          type: integer
          description: 'Cantidad de créditos.'
          examples:
            - 4
      required:
        - id
        - usuario_id
        - nombre
        - codigo
        - semestre
        - creditos
      title: CursoResource
    PendienteResource:
      type: object
      properties:
        id:
          type: integer
          description: 'Identificador del pendiente.'
          examples:
            - 1
        usuario_id:
          type: integer
          description: 'Identificador del usuario propietario.'
          examples:
            - 1
        curso_id:
          type:
            - integer
            - 'null'
          description: 'Identificador del curso asociado.'
          examples:
            - 1
        titulo:
          type: string
          description: 'Título del pendiente.'
          examples:
            - 'Entregar proyecto de Desarrollo de Software'
        descripcion:
          type:
            - string
            - 'null'
          description: 'Descripción del pendiente.'
          examples:
            - 'Completar y entregar el proyecto final del curso.'
        estado:
          type: string
          description: 'Estado actual del pendiente.'
          examples:
            - pendiente
        fecha_limite:
          type:
            - string
            - 'null'
          description: 'Fecha límite.'
          examples:
            - '2026-10-15'
        hora_pendiente:
          type:
            - string
            - 'null'
          description: 'Hora asociada al pendiente.'
          examples:
            - '18:00:00'
      required:
        - id
        - usuario_id
        - curso_id
        - titulo
        - descripcion
        - estado
        - fecha_limite
        - hora_pendiente
      title: PendienteResource
...
    StorePendienteConRecordatorioRequest:
      type: object
      properties:
        usuario_id:
          type: integer
        curso_id:
          type:
            - integer
            - 'null'
        titulo:
          type: string
          maxLength: 150
        descripcion:
          type:
            - string
            - 'null'
          maxLength: 1000
        estado:
          type: string
          enum:
            - pendiente
            - en_progreso
            - completado
        fecha_limite:
          type: string
          format: date-time
        hora_pendiente:
          type:
            - string
            - 'null'
        recordatorio:
          type: object
          properties:
            titulo:
              type: string
              maxLength: 150
            descripcion:
              type:
                - string
                - 'null'
              maxLength: 1000
            fecha_recordatorio:
              type: string
              format: date-time
            hora_inicio:
              type: string
          required:
            - titulo
            - fecha_recordatorio
            - hora_inicio
      required:
        - usuario_id
        - titulo
        - estado
        - fecha_limite
        - recordatorio
      title: StorePendienteConRecordatorioRequest
...
    UpdateCursoRequest:
      type: object
      properties:
        usuario_id:
          type: integer
        nombre:
          type: string
          maxLength: 100
        codigo:
          type: string
          maxLength: 20
        semestre:
          type: string
          maxLength: 20
        creditos:
          type: integer
          minimum: 1
          maximum: 20
      title: UpdateCursoRequest
    UpdatePendienteRequest:
      type: object
      properties:
        usuario_id:
          type: integer
        curso_id:
          type:
            - integer
            - 'null'
        titulo:
          type: string
          maxLength: 150
        descripcion:
          type:
            - string
            - 'null'
          maxLength: 1000
        estado:
          type: string
          enum:
            - pendiente
            - en_progreso
            - completado
        fecha_limite:
          type:
            - string
            - 'null'
          format: date-time
        hora_pendiente:
          type:
            - string
            - 'null'
      title: UpdatePendienteRequest
  responses:
    ValidationException:
      description: 'Validation error'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Errors overview.'
              errors:
                type: object
                description: 'A detailed description of each field that failed validation.'
                additionalProperties:
                  type: array
                  items:
                    type: string
            required:
              - message
              - errors
    ModelNotFoundException:
      description: 'Not found'
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: 'Error overview.'
            required:
              - message